requirements.txt
python-dotenv
psycopg2-binary
streamlit
pandas
fpdf2